EmployeeSalary/src/Lab01_22H4060068/SalariedEmployee.java
package Lab01_22H4060068;

public class SalariedEmployee extends Employee {
    private double weeklySalary;

    // constructor
    public SalariedEmployee(String firstName, String lastName, String socialSecurityNumber, double weeklySalary) {
        super(firstName, lastName, socialSecurityNumber);
        if (weeklySalary < 0.0) {
            throw new IllegalArgumentException("Weekly salary must be >= 0.0");
        }
        this.weeklySalary = weeklySalary;
    }

    // get weekly salary
    public double getWeeklySalary() {
        return weeklySalary;
    }

    // set weekly salary
    public void setWeeklySalary(double weeklySalary) {
        if (weeklySalary < 0.0) {
            throw new IllegalArgumentException("Weekly salary must be >= 0.0");
        }
        this.weeklySalary = weeklySalary;
    }

    // calculate earnings
    @Override
    public double earnings() {
        return getWeeklySalary();
    }

    // return String representation
    @Override
    public String toString() {
        return String.format("salaried employee: %s\n%s: $%,.2f", super.toString(), "weekly salary", getWeeklySalary());
    }
}
